requirements.txt
aioredis==1.3.1
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.4.5.2
cffi==1.14.1
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.0
daphne==2.5.0
DateTime==4.3
DBUtils==1.3
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0.6
django-cors-headers==3.3.0
django-import-export==2.3.0
djangorestframework==3.11.1
dwebsocket==0.5.12
et-xmlfile==1.0.1
future==0.18.2
geocoder==1.38.1
geographiclib==1.50
geopy==1.22.0
gevent==20.9.0
greenlet==0.4.17
hiredis==1.1.0
hyperlink==20.0.1
idna==2.9
incremental==17.5.0
jdcal==1.4.1
MarkupPy==1.14
msgpack==1.0.0
numpy==1.19.1
odfpy==1.4.1
openpyxl==3.0.4
pandas==1.1.1
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
ratelim==0.1.6
redis==3.5.3
requests==2.24.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
treelib==1.6.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
uWSGI==2.0.19.1
websocket==0.2.1
xlrd==1.2.0
xlwt==1.3.0
zope.event==4.5.0
zope.interface==5.1.0